Address custom sidebar reload review feedback
